Add documentation
Shorten and embiggen the app name at the top of the screen, and add a
`info.circle` ⓘ button to the right.  Tapping the ⓘ button will
show brief usage docs in a sheet, which can be dismissed by tapping
the Dismiss button.

Update the **Matches** section to print "No matching words found" if
there are no matches, and if all textboxes are empty, *also* show
"Try entering some letters?"